poe_mover/requirements.txt
anthropic>=0.25.0
Pillow>=10.0.0
pyautogui>=0.9.54
pygetwindow>=0.0.9; sys_platform == "win32"
